Copied Codepath's template over to provide outline.
https://courses.codepath.com/snippets/cybersecurity_university/readme_templates/project_7_readme.md?raw=true
Be sure to include a README.md based on this README template containing a small writeup and GIF walkthroughs of each exploit.
This assignment doesn't require submission of additional code, but you are encouraged to include any code, scripts, or assets you used in your repository. Note the README and associated GIFs are required.